refactor: move table related types into tables folder.
