+++ src/app/projects/[projectId]/_components/edit-project/EditProjectForm.tsx
@@ -3,7 +3,7 @@
 import { useDialog } from '@/app/projects/[projectId]/_components/dialog-context';
 import ProjectForm from '@/components/shared/forms/ProjectForm';
 import { editProject } from '@/lib/actions/project/edit-project';
-import { TProject } from '@/lib/types/project';
+import { TProject } from '@/lib/types/tables/project';
 
 type TEditProjectFormProps = {
   project: TProject;
