+++ src/lib/actions/auth/sign-in.ts
@@ -7,7 +7,7 @@ import { users } from '@/drizzle/schema';
 import { db } from '@/lib/db';
 import { createSession } from '@/lib/server/session';
 import { TFormActionReturn } from '@/lib/types/form-action-return';
-import { DB_USER_INCLUDED_COLUMNS, TUser } from '@/lib/types/user';
+import { DB_USER_INCLUDED_COLUMNS, TUser } from '@/lib/types/tables/user';
 import { TZodObjectErrors } from '@/lib/types/zod-object-errors';
 import { actionError } from '@/lib/utils/action-error';
 import {
