+++ src/lib/server/get-user.ts
@@ -3,7 +3,7 @@ import { cache } from 'react';
 import { users } from '@/drizzle/schema';
 import { db } from '@/lib/db';
 import { getSession, verifySession } from '@/lib/server/session';
-import { DB_USER_INCLUDED_COLUMNS, TUser } from '@/lib/types/user';
+import { DB_USER_INCLUDED_COLUMNS, TUser } from '@/lib/types/tables/user';
 
 async function fetchUserFromDb(userId: string): Promise<TUser> {
   const user: TUser | undefined = await db.query.users.findFirst({
