+++ src/lib/types/tables/project.ts
@@ -1,7 +1,7 @@
 import { InferSelectModel } from 'drizzle-orm';
 import { projectMembers, projects } from '@/drizzle/schema';
-import { TTaskWithMemberIds } from '@/lib/types/task';
-import { TUser } from '@/lib/types/user';
+import { TTaskWithMemberIds } from '@/lib/types/tables/task';
+import { TUser } from '@/lib/types/tables/user';
 
 export type TProject = InferSelectModel<typeof projects>;
 export type TProjectWithTasksProgress = TProject & {
